requirements.txt
faiss-cpu == 1.8.0
sentence-transformers == 2.6.1
Flask == 3.0.3
numpy == 1.26.4
pandas == 2.0.3
torch == 2.3.0
transformers == 4.40.2
google-cloud-storage == 2.16.0
gunicorn == 22.0.0